Issue-9704 Put python check after -r and -d checks
When they python interpreter is set incorrectly for the machine the file
is being checked for (e.g. for the local or the remote), the error
manifests as a readability or directory missing error which can be very
misleading. By returning before checking for the python interpreter we
can cause it to fail with a proper error message.